# More restore optimizations