--- a/corehq/ex-submodules/casexml/apps/phone/models.py
+++ b/corehq/ex-submodules/casexml/apps/phone/models.py
@@ -759,16 +759,6 @@ class SimplifiedSyncLog(AbstractSyncLog):
             self._purged_cases = set()
         return self._purged_cases
 
-    _frozen_closed_cases = None
-
-    @property
-    def frozen_closed_cases(self):
-        """Used so that we can memoize a function that requires this
-        """
-        if self._frozen_closed_cases is None:
-            self._frozen_closed_cases = frozenset(self.closed_cases)
-        return self._frozen_closed_cases
-
     def save(self, *args, **kwargs):
         # force doc type to SyncLog to avoid changing the couch view.
         self.doc_type = "SyncLog"
@@ -894,7 +884,7 @@ class SimplifiedSyncLog(AbstractSyncLog):
             new_live = new_live | IndexTree.traverse_incoming_extensions(
                 case_to_check,
                 self.extension_index_tree,
-                self.frozen_closed_cases,
+                frozenset(self.closed_cases),
             ) - self.purged_cases
             new_live = new_live - checked
             live = live | new_live